sigstore, test: more unit tests, rename mods
