+++ sigstore/_internal/rekor/__init__.py
@@ -16,6 +16,6 @@
 APIs for interacting with Rekor.
 """
 
-from ._client import RekorClient, RekorEntry, RekorInclusionProof
+from .client import RekorClient, RekorEntry, RekorInclusionProof
 
 __all__ = ["RekorClient", "RekorEntry", "RekorInclusionProof"]
