+++ test/internal/rekor/__init__.py
@@ -11,22 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-"""
-APIs for interacting with Fulcio.
-"""
-
-
-from .client import (
-    FulcioCertificateSigningRequest,
-    FulcioCertificateSigningResponse,
-    FulcioClient,
-    FulcioSignedCertificateTimestamp,
-)
-
-__all__ = [
-    "FulcioCertificateSigningRequest",
-    "FulcioCertificateSigningResponse",
-    "FulcioClient",
-    "FulcioSignedCertificateTimestamp",
-]
